refactor(controllers): mouve getEditModerationController in user
