+++ src/controllers/user/update-personal-informations.ts
@@ -1,5 +1,8 @@
 import type { NextFunction, Request, Response } from "express";
+import HttpErrors from "http-errors";
 import { z, ZodError } from "zod";
+import { NotFoundError } from "../../config/errors";
+import { getOrganizationFromModeration } from "../../managers/moderation";
 import {
   getUserFromAuthenticatedSession,
   updateUserInAuthenticatedSession,
@@ -7,6 +10,7 @@ import {
 import { updatePersonalInformations } from "../../managers/user";
 import { csrfToken } from "../../middlewares/csrf-protection";
 import {
+  idSchema,
   jobSchema,
   nameSchema,
   phoneNumberSchema,
@@ -81,6 +85,38 @@ export const postPersonalInformationsController = async (
       );
     }
 
+    next(error);
+  }
+};
+export const getEditModerationController = async (
+  req: Request,
+  res: Response,
+  next: NextFunction,
+) => {
+  try {
+    const user = getUserFromAuthenticatedSession(req);
+
+    const schema = z.object({
+      moderation_id: idSchema(),
+    });
+    let { moderation_id } = await schema.parseAsync(req.query);
+
+    const { cached_libelle } = await getOrganizationFromModeration({
+      user,
+      moderation_id,
+    });
+
+    return res.render("edit-moderation", {
+      pageTitle: "Modifier une demande de rattachement en cours",
+      email: user.email,
+      csrfToken: user.email && csrfToken(req),
+      organization_label: cached_libelle,
+      moderation_id,
+    });
+  } catch (error) {
+    if (error instanceof NotFoundError) {
+      return next(new HttpErrors.NotFound());
+    }
     next(error);
   }
 };
